% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive.R
\name{consecutive}
\alias{consecutive}
\title{Number of consecutive occurences}
\usage{
consecutive(x_lgl, show_where = FALSE)
}
\arguments{
\item{x_lgl}{logical vector. It indicates the condition for which the number of consective occurences is calculated.}

\item{show_where}{logical. It reports starting and ending positions of consecutive pattern.}
}
\value{
a numerical vector or a list. If show_where = FALSE the function returns only the lengths of consecutive occurences. If show_where = TRUE the function returns a list containg the lengths of consecutive occurences as well as their starting and ending postiions.
}
\description{
Number of consecutive occurences
}
\examples{
x <- c(1:3, rep(0,4), 5:7, rep(0,6), 9:10)
consecutive(x == 0)
consecutive(x == 0, show_where = TRUE)
}
