% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embrace_get.R
\name{embrace_get}
\alias{embrace_get}
\title{Get string between embrace}
\usage{
embrace_get(.x, .left = "\\\\(", .right = "\\\\)")
}
\arguments{
\item{.right}{}
}
\description{
Get string between embrace
}
