% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list}
\alias{list}
\title{Named lists}
\usage{
list(..., .named = TRUE, .base = FALSE)
}
\arguments{
\item{...}{Objects to include in the list.}
}
\value{
A named list.
}
\description{
Create a named list using specified names or, if names are omitted, using the names of the objects in the list. The code list(a = a, b = b) becomes nlist(a,b) and list(a = a, b = 2) becomes nlist(a, b = 2), etc.
}
