% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{Mini-max - coerce to (0/1) or (0/100)}
\usage{
minmax(x, min, max, x100 = FALSE, alpha = 0)
}
\arguments{
\item{alpha}{}
}
\description{
Mini-max - coerce to (0/1) or (0/100)
}
