% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_summary.R
\name{n_missing}
\alias{n_missing}
\title{Summary of Missing Values}
\usage{
n_missing(x)
}
\arguments{
\item{x}{a vector.}
}
\value{
a number.
}
\description{
Summary of Missing Values
}
\examples{
x <- c(1,2,3,NA,5,NA,7,8,9,10)
n_missing(x)
n_complete(x)
missing_pct(x)
complete_pct(x)

}
