% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_in.R
\name{names_in}
\alias{names_in}
\title{Names of Vector, Table in a character vector}
\usage{
names_in(x, values, negate = FALSE, which = FALSE)
}
\arguments{
\item{which}{}
}
\description{
Names of Vector, Table in a character vector
}
