% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numpad.R
\name{numpad}
\alias{numpad}
\title{Padding numeric values}
\usage{
numpad(x, pad = "0")
}
\arguments{
\item{x}{an integer or a vector of integers.}

\item{pad}{a character input indicating what is padding the input.}
}
\value{
a string or character vector of padded values.
}
\description{
Padding numeric values
}
\examples{
numpad(1:15)


}
