% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_upper_split.R
\name{str_upper_split}
\alias{str_upper_split}
\title{Split String at Uppercse}
\usage{
str_upper_split(string, sep = "_", to_lower = TRUE)
}
\arguments{
\item{string}{the string to modify}

\item{sep}{a string indicating the separator.}

\item{to_lower}{TRUE or FALSE}
}
\value{
a string
}
\description{
Split String at Uppercse
}
