% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{Mini-max - coerce to (0/1) or (0/100)}
\usage{
minmax(x, min = NULL, max = NULL, to_100 = FALSE)
}
\arguments{
\item{to_100}{}
}
\description{
Mini-max - coerce to (0/1) or (0/100)
}
